refactor: middleware and added types for update extension request
